Move RewardsTransactionCache to components/staking
